+++ src/test/java/nextstep/subway/acceptance/AcceptanceTestSteps.java
@@ -5,12 +5,12 @@ import io.restassured.specification.RequestSpecification;
 
 public class AcceptanceTestSteps {
 
-    static RequestSpecification given() {
+    public static RequestSpecification given() {
         return RestAssured
                 .given().log().all();
     }
 
-    static RequestSpecification given(String token) {
+    public static RequestSpecification given(String token) {
         return RestAssured
                 .given().log().all()
                 .auth().oauth2(token);
