+++ src/test/java/nextstep/subway/acceptance/member/MemberSteps.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.member;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
