+++ src/test/java/nextstep/subway/acceptance/path/PathSteps.java
@@ -1,8 +1,9 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.path;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
+import nextstep.subway.acceptance.AcceptanceTestSteps;
 import org.springframework.http.MediaType;
 
 import static nextstep.subway.acceptance.AcceptanceTestSteps.given;
@@ -16,7 +17,7 @@ public class PathSteps {
     }
 
     public static ExtractableResponse<Response> 두_역의_최단_거리_경로_조회를_요청(Long source, Long target) {
-        return given()
+        return AcceptanceTestSteps.given()
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/paths?source={sourceId}&target={targetId}", source, target)
                 .then().log().all().extract();
